refactor(auth-home-register): moduleize auth social flow and legacy screens
