+++ src/screens/auth/index.ts
@@ -1,0 +1,4 @@
+export { default as LoginScreen } from './login';
+export { default as RegisterScreen } from './register';
+export { SocialLoginScreen } from './social';
+export * from './routes';
