+++ src/screens/auth/routes.ts
@@ -1,0 +1,5 @@
+export const AUTH_ROUTES = {
+  LOGIN: 'Login',
+  REGISTER: 'Register',
+  SOCIAL_LOGIN: 'SocialLogin',
+} as const;
