+++ src/screens/auth/social/index.ts
@@ -1,0 +1,1 @@
+export { SocialLoginScreen } from './view/SocialLoginScreen';
