+++ src/screens/home/home/index.ts
@@ -1,0 +1,1 @@
+export { HomeScreen } from './view/HomeScreen';
