+++ src/screens/home/index.ts
@@ -1,0 +1,2 @@
+export { HomeScreen } from './home';
+export * from './routes';
