+++ src/screens/home/routes.ts
@@ -1,0 +1,3 @@
+export const HOME_ROUTES = {
+  HOME: 'Home',
+} as const;
