+++ src/screens/register/home/index.ts
@@ -1,0 +1,1 @@
+export { RegisterScreen } from './view/RegisterScreen';
