+++ src/screens/register/index.ts
@@ -1,0 +1,2 @@
+export { RegisterScreen } from './home';
+export * from './routes';
