+++ src/screens/register/routes.ts
@@ -1,0 +1,3 @@
+export const REGISTER_ROUTES = {
+  HOME: 'Register',
+} as const;
